refactor: OrderStatus order로 이동 및 validator 책임 분리
