+++ src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,7 +1,6 @@
 package kitchenpos.order.domain;
 
 import java.util.List;
-import kitchenpos.common.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
